refactor: refactoring ClaimRewards file names and fixing button texts
